+++ HardAntiCheat/bin/Debug/netstandard2.1/_ThunderstoreBuild/CHANGELOG.md
@@ -9,7 +9,9 @@
 - Ported Codetalker varients - 10/14/25
 - Extended verification length - 10/14/25
 - Anticheat Integ Check Addition to minimize chances of bypassing - 10/17/25
-- Game update [+] 10/21/25
+- Game update
+- Added Mod Validation/Configurations (LIVE TESTING) [+] 10/29/25
+- Patched FluffUtilities (LIVE TESTING) [+] 10/29/25
 
 ## V2 - Refactory - Public Test Update (cont ^)
 - Fixes "Players getting railed on skill init still" FULLY